+++ tests/cluster/go-test/filter-oauth2/testdata/idp_auth0.js
@@ -1,20 +1,19 @@
-module.exports.testcases = {
-	"Auth0 (/auth0/httpbin)": {
-		resource: "https://ambassador.standalone.svc.cluster.local/auth0/httpbin/headers",
-		username: "testuser@datawire.com",
-		password: "TestUser321",
-	},
-	"Auth0 (/auth0-k8s/httpbin)": {
-		resource: "https://ambassador.standalone.svc.cluster.local/auth0-k8s/httpbin/headers",
-		username: "testuser@datawire.com",
-		password: "TestUser321",
-	},
-	"Auth0 (/httpbin)": {
-		resource: "https://ambassador.standalone.svc.cluster.local/httpbin/headers",
-		username: "testuser@datawire.com",
-		password: "TestUser321",
-	}
-};
+module.exports.testcases = (() => {
+	let testcases = {}
+	let addTestcase = function(name) {
+		testcases[`Auth0 (/${name})`] = {
+			resource: `https://ambassador.standalone.svc.cluster.local/${name}/headers`,
+			username: "testuser@datawire.com",
+			password: "TestUser321",
+		};
+	};
+	addTestcase('oauth2-auth0-nojwt-and-plugin-and-whitelist');
+	addTestcase('oauth2-auth0-nojwt-and-k8ssecret-and-xhrerror');
+	addTestcase('oauth2-auth0-nojwt-and-anyerror');
+	addTestcase('oauth2-auth0-simplejwt');
+	addTestcase('oauth2-auth0-complexjwt');
+	return testcases;
+})();
 
 module.exports.authenticate = async function(browsertab, username, password) {
 	console.log("[auth0] email...");